FLNotify/requirements.txt
flask
boto3
python-dotenv
numpy
scikit-learn
